Create amirNobari.js
In this code, a function called findMostFrequentElement is written, which is responsible for finding the element that has the most repetition in a given array.

Contextual functions (Arrow Functions) in JavaScript are briefly functions that can be easily written in one line. Here, a context function is used to define the function.

The findMostFrequentElement function takes an array called arr as input. Then, using the reduce function, it goes through this array and follows the most repeated.

In the reduce function, the variable a represents the aggregation state that is initially initialized with the value { element: null }. The variable c represents the current element in the array and i represents the index of the current element. t also refers to an array.

In each step of reduce, the value of a can be compared with the current value. For this purpose, the filter function is used for the t array. This filter function selects elements that are equal to c (the current element) and returns the number of occurrences of that element.

If the number of occurrences of the current element (c) is greater than the number of occurrences of the element in a.element, a is set to the current element (c). Finally, the element with the most repetition in the array is returned as the result.

This code is able to find the most repetition in a very short way and returns the element corresponding to that repetition as a result.